requirements.txt
Django==1.10.2
argparse==1.2.1
django-allauth==0.27.0
djangorestframework==3.4.7
oauthlib==2.0.0
python-openid==2.2.5
requests==2.11.1
requests-oauthlib==0.7.0
wsgiref==0.1.2
